Move gutter mark into generated highlights
