+++ resharper/src/resharper-unity/Daemon/Stages/GutterMarks/UnityFieldDetector.cs
@@ -1,13 +1,11 @@
-using JetBrains.ReSharper.Daemon.Stages.Dispatcher;
 using JetBrains.ReSharper.Feature.Services.Daemon;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Errors;
 using JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Dispatcher;
-using JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Highlightings;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
-using JetBrains.ReSharper.Psi.Tree;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.GutterMarks
 {
-    [ElementProblemAnalyzer(typeof(IFieldDeclaration), HighlightingTypes = new[] { typeof(UnityMarkOnGutter) })]
+    [ElementProblemAnalyzer(typeof(IFieldDeclaration), HighlightingTypes = new[] { typeof(UnityGutterMarkInfo) })]
     public class UnityFieldDetector : UnityElementProblemAnalyzer<IFieldDeclaration>
     {
         public UnityFieldDetector(UnityApi unityApi)
@@ -19,7 +17,10 @@ namespace JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.GutterMarks
         {
             var field = element.DeclaredElement;
             if (field != null && Api.IsUnityField(field))
-                AddGutterMark(element, element.GetNameDocumentRange(), "This field is initialised by Unity", consumer);
+            {
+                var highlighting = new UnityGutterMarkInfo(element, "This field is initialised by Unity");
+                consumer.AddHighlighting(highlighting);
+            }
         }
     }
 }
