+++ resharper/test/src/Daemon/Stages/GutterMarks/GutterMarkTests.cs
@@ -1,6 +1,6 @@
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.FeaturesTestFramework.Daemon;
-using JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Highlightings;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Errors;
 using JetBrains.ReSharper.Psi;
 using NUnit.Framework;
 
@@ -13,7 +13,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Tests.Daemon.Stages.GutterMarks
 
         protected override bool HighlightingPredicate(IHighlighting highlighting, IPsiSourceFile sourceFile)
         {
-            return highlighting is UnityMarkOnGutter;
+            return highlighting is UnityGutterMarkInfo;
         }
 
         [Test] public void Test01() { DoNamedTest(); }
